Refactor call handling
