--- proc-macro/src/assert_pre.rs
@@ -1,412 +1,0 @@
-//! Functionality for parsing and visiting `assert_pre` attributes.
-
-use proc_macro2::Span;
-use proc_macro_error::{emit_error, emit_warning};
-use quote::{quote, quote_spanned};
-use std::mem;
-use syn::{
-    parenthesized,
-    parse::{Parse, ParseStream},
-    parse2,
-    punctuated::Pair,
-    spanned::Spanned,
-    token::Paren,
-    Expr, ExprPath, LitStr, Path, Token,
-};
-
-use crate::{
-    call::Call,
-    helpers::{visit_matching_attrs_parsed, Parenthesized},
-    precondition::Precondition,
-    render_assert_pre,
-};
-
-/// The custom keywords used in the `assert_pre` attribute.
-mod custom_keywords {
-    use syn::custom_keyword;
-
-    custom_keyword!(def);
-    custom_keyword!(reason);
-}
-
-/// An `assert_pre` declaration.
-enum AssertPreAttr {
-    /// Information where to find the definition of the preconditions.
-    DefStatement(Parenthesized<DefStatement>),
-    /// A statement that the precondition holds.
-    Precondition(Parenthesized<PreconditionHoldsStatement>),
-}
-
-impl Parse for AssertPreAttr {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let content;
-        let parentheses = parenthesized!(content in input);
-
-        Ok(if content.peek(custom_keywords::def) {
-            AssertPreAttr::DefStatement(Parenthesized::with_parentheses(parentheses, &content)?)
-        } else {
-            AssertPreAttr::Precondition(Parenthesized::with_parentheses(parentheses, &content)?)
-        })
-    }
-}
-
-/// A statement that a precondition holds.
-enum PreconditionHoldsStatement {
-    /// The statement had a reason attached to it.
-    WithReason {
-        /// The precondition that was stated.
-        precondition: Precondition,
-        /// The comma separating the precondition from the reason.
-        _comma: Token![,],
-        /// The reason that was stated.
-        reason: Reason,
-    },
-    /// The statement written without a reason.
-    WithoutReason {
-        /// The precondition that was stated.
-        precondition: Precondition,
-        /// The span where to place the missing reason.
-        missing_reason_span: Span,
-    },
-}
-
-impl PreconditionHoldsStatement {
-    /// Returns only the precondition from this statement.
-    fn into_precondition(self) -> Precondition {
-        match self {
-            PreconditionHoldsStatement::WithoutReason { precondition, .. } => precondition,
-            PreconditionHoldsStatement::WithReason { precondition, .. } => precondition,
-        }
-    }
-}
-
-impl Parse for PreconditionHoldsStatement {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let precondition = input.parse()?;
-
-        if input.is_empty() {
-            Ok(PreconditionHoldsStatement::WithoutReason {
-                precondition,
-                missing_reason_span: input.span(),
-            })
-        } else {
-            let comma = input.parse()?;
-            let reason = input.parse()?;
-
-            Ok(PreconditionHoldsStatement::WithReason {
-                precondition,
-                _comma: comma,
-                reason,
-            })
-        }
-    }
-}
-
-/// The reason why a precondition holds.
-struct Reason {
-    /// The `reason` keyword.
-    _reason_keyword: custom_keywords::reason,
-    /// The `=` separating the `reason` keyword and the reason.
-    _eq: Token![=],
-    /// The reason the precondition holds.
-    reason: LitStr,
-}
-
-impl Parse for Reason {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let reason_keyword = input.parse()?;
-        let eq = input.parse()?;
-        let reason = input.parse()?;
-
-        Ok(Reason {
-            _reason_keyword: reason_keyword,
-            _eq: eq,
-            reason,
-        })
-    }
-}
-
-/// Provides information where to find the definition of the preconditions.
-struct DefStatement {
-    /// The def keyword.
-    def_keyword: custom_keywords::def,
-    /// The parentheses surrounding the definition site.
-    parentheses: Paren,
-    /// Information about the definition site.
-    site: DefStatementSite,
-}
-
-impl Parse for DefStatement {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let def_keyword = input.parse()?;
-        let content;
-        let parentheses = parenthesized!(content in input);
-        let site = content.parse()?;
-
-        Ok(DefStatement {
-            def_keyword,
-            parentheses,
-            site,
-        })
-    }
-}
-
-impl Spanned for DefStatement {
-    fn span(&self) -> Span {
-        self.def_keyword
-            .span()
-            .join(self.parentheses.span)
-            .unwrap_or_else(|| self.parentheses.span)
-    }
-}
-
-impl DefStatement {
-    /// Constructs a new path correctly using the new definition site.
-    fn construct_new_path(&self, fn_path: &ExprPath) -> ExprPath {
-        let mut resulting_path = fn_path.clone();
-
-        match self.site {
-            DefStatementSite::Direct { ref path } => {
-                for (i, segment) in path.segments.iter().enumerate() {
-                    resulting_path.path.segments.insert(i, segment.clone());
-                }
-            }
-            DefStatementSite::Replace {
-                ref from, ref to, ..
-            } => {
-                if from.segments.len() > resulting_path.path.segments.len() {
-                    emit_error!(
-                        fn_path,
-                        "cannot replace `{}` in this path",
-                        quote! { #from };
-                        help = from.span()=> "try specifing a prefix of `{}` in `def(...)`",
-                        quote! { #fn_path }
-                    );
-                    return resulting_path;
-                }
-
-                for (from_segment, fn_segment) in from
-                    .segments
-                    .iter()
-                    .zip(resulting_path.path.segments.iter())
-                {
-                    if from_segment != fn_segment {
-                        emit_error!(
-                            fn_path,
-                            "cannot replace `{}` in this path",
-                            quote! { #from };
-                            note = fn_segment.span()=> "`{}` != `{}`",
-                            quote! { #from_segment },
-                            quote! { #fn_segment };
-                            help = from.span()=> "try specifing a prefix of `{}` in `def(...)`",
-                            quote! { #fn_path }
-                        );
-                        return resulting_path.clone();
-                    }
-                }
-
-                resulting_path.path.segments = to
-                    .segments
-                    .pairs()
-                    .map(|pair| match pair {
-                        Pair::Punctuated(segment, punct) => {
-                            Pair::Punctuated(segment.clone(), *punct)
-                        }
-                        Pair::End(segment) => {
-                            if resulting_path.path.segments.len() > from.segments.len() {
-                                // If there is more path to come, don't put the end into the
-                                // iterator yet. The next iterator will take care of the end.
-                                Pair::Punctuated(segment.clone(), Default::default())
-                            } else {
-                                Pair::End(segment.clone())
-                            }
-                        }
-                    })
-                    .chain(
-                        resulting_path
-                            .path
-                            .segments
-                            .pairs()
-                            .map(|pair| match pair {
-                                Pair::Punctuated(segment, punct) => {
-                                    Pair::Punctuated(segment.clone(), *punct)
-                                }
-                                Pair::End(segment) => Pair::End(segment.clone()),
-                            })
-                            .skip(from.segments.len()),
-                    )
-                    .collect();
-            }
-        }
-
-        resulting_path
-    }
-}
-
-/// Provides the definition in a `def(...)` statement.
-enum DefStatementSite {
-    /// The definition is found directly at the given path.
-    Direct {
-        /// The path where to find the definition.
-        path: Path,
-    },
-    /// The definition is found by replacing `from` with `to` in the path.
-    Replace {
-        /// The path where the original function is found.
-        from: Path,
-        /// The arrow token that marks the replacement.
-        _arrow: Token![->],
-        /// The path where to function with the attached preconditions is found.
-        to: Path,
-    },
-}
-
-impl Parse for DefStatementSite {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let first_path = input.parse()?;
-
-        Ok(if input.is_empty() {
-            DefStatementSite::Direct { path: first_path }
-        } else {
-            let arrow = input.parse()?;
-            let second_path = input.parse()?;
-
-            DefStatementSite::Replace {
-                from: first_path,
-                _arrow: arrow,
-                to: second_path,
-            }
-        })
-    }
-}
-
-/// The reason to display in the hint where to add the reason.
-const HINT_REASON: &str = "why does this hold?";
-
-/// The name of the macro used to assert that a condition holds.
-const ASSERT_CONDITION_HOLDS_ATTR: &str = "assert_pre";
-
-/// Renders the expression
-pub(crate) fn process_call(mut call: Call) -> Option<Expr> {
-    let mut def_statement = None;
-    let mut preconditions = Vec::new();
-
-    let attr_span = visit_matching_attrs_parsed(
-        call.attrs_mut(),
-        |attr| attr.path.is_ident(ASSERT_CONDITION_HOLDS_ATTR),
-        |parsed_attr| match parsed_attr {
-            AssertPreAttr::DefStatement(Parenthesized { content: def, .. }) => {
-                if let Some(old_def_statement) = def_statement.replace(def) {
-                    let span = def_statement
-                        .as_ref()
-                        .expect("options contains a value, because it was just put there")
-                        .span();
-                    emit_error!(
-                        span,
-                        "duplicate `def(...)` statement";
-                        help = old_def_statement.span() => "there can be just one definition site, try removing the wrong one"
-                    );
-                }
-            }
-            AssertPreAttr::Precondition(Parenthesized {
-                content: precondition,
-                ..
-            }) => {
-                preconditions.push(precondition);
-            }
-        },
-    );
-
-    if let Some(attr_span) = attr_span {
-        Some(render_call(preconditions, def_statement, attr_span, call))
-    } else {
-        None
-    }
-}
-
-/// Returns an unfinished reason declaration for the precondition if one exists.
-fn unfinished_reason(reason: &LitStr) -> Option<&LitStr> {
-    let mut reason_val = reason.value();
-
-    reason_val.make_ascii_lowercase();
-    match &*reason_val {
-        HINT_REASON | "todo" | "?" => Some(reason),
-        _ => None,
-    }
-}
-
-/// Process a found `assert_pre` attribute.
-fn render_call(
-    preconditions: Vec<PreconditionHoldsStatement>,
-    def_statement: Option<DefStatement>,
-    attr_span: Span,
-    mut call: Call,
-) -> Expr {
-    for precondition in preconditions.iter() {
-        match precondition {
-            PreconditionHoldsStatement::WithReason { reason, .. } => {
-                if let Some(reason) = unfinished_reason(&reason.reason) {
-                    emit_warning!(
-                        reason,
-                        "you should specify a more meaningful reason here";
-                        help = "specifying a meaningful reason here will help you and others understand why this is ok in the future"
-                    )
-                }
-            }
-            PreconditionHoldsStatement::WithoutReason {
-                precondition,
-                missing_reason_span,
-                ..
-            } => emit_error!(
-                precondition.span(),
-                "you need to specify a reason why this precondition holds";
-                help = *missing_reason_span => "add `, reason = {:?}`", HINT_REASON
-            ),
-        }
-    }
-
-    let preconditions = preconditions
-        .into_iter()
-        .map(|holds_statement| holds_statement.into_precondition())
-        .collect();
-
-    let mut original_call = None;
-
-    if let Some(def_statement) = def_statement {
-        original_call = Some(call.clone());
-
-        match &mut call {
-            Call::Function(ref mut call) => {
-                if let Expr::Path(p) = *call.func.clone() {
-                    let mut new_path = Expr::Path(def_statement.construct_new_path(&p));
-
-                    mem::swap(&mut *call.func, &mut new_path);
-                } else {
-                    emit_error!(
-                        call.func,
-                        "unable to determine at compile time which function is being called";
-                        help = "use a direct path to the function instead"
-                    );
-                }
-            }
-            _ => todo!(),
-        }
-    }
-
-    let output = render_assert_pre(preconditions, call, attr_span);
-
-    if let Some(original_call) = original_call {
-        parse2(quote_spanned! {
-            original_call.span()=>
-                #[allow(dead_code)]
-                if true {
-                    #output
-                } else {
-                    #original_call
-                }
-        })
-        .expect("if expression is a valid expression")
-    } else {
-        output.into()
-    }
-}
