+++ proc-macro/src/helpers.rs
@@ -64,7 +64,7 @@ pub(crate) fn visit_matching_attrs_parsed<ParsedAttr: Parse>(
 /// A parsable thing surrounded by parentheses.
 pub(crate) struct Parenthesized<T> {
     /// The parentheses surrounding the object.
-    _parentheses: Paren,
+    pub(crate) parentheses: Paren,
     /// The content that was surrounded by the parentheses.
     pub(crate) content: T,
 }
@@ -73,7 +73,7 @@ impl<T: Parse> Parenthesized<T> {
     /// Parses the content, if the parentheses were already parsed.
     pub(crate) fn with_parentheses(parentheses: Paren, input: ParseStream) -> syn::Result<Self> {
         Ok(Parenthesized {
-            _parentheses: parentheses,
+            parentheses,
             content: input.parse()?,
         })
     }
@@ -86,7 +86,7 @@ impl<T: Parse> Parse for Parenthesized<T> {
         let content = content.parse()?;
 
         Ok(Parenthesized {
-            _parentheses: parentheses,
+            parentheses,
             content,
         })
     }
