+++ proc-macro/src/lib.rs
@@ -13,8 +13,8 @@ use crate::{
     pre_defs_for::{DefinitionsForAttr, DefinitionsForModule},
 };
 
-mod assert_pre;
 mod call;
+mod call_handling;
 mod helpers;
 mod pre_attr;
 mod pre_defs_for;
