+++ proc-macro/src/pre_attr.rs
@@ -13,8 +13,8 @@ use syn::{
 };
 
 use crate::{
-    assert_pre::process_call,
     call::Call,
+    call_handling::process_call,
     helpers::{crate_name, visit_matching_attrs_parsed, Parenthesized},
     precondition::Precondition,
     render_pre,
